Add points listing API
