+++ src/models/model_base.py
@@ -8,6 +8,18 @@ class ModelBase(db.Model):
     created_on = db.Column(db.DateTime, server_default=db.func.now())
     updated_on = db.Column(db.DateTime, server_default=db.func.now(), onupdate=db.func.now())
 
+    @classmethod
+    def find_all(cls):
+        return cls.query.all()
+
+    @classmethod
+    def filter_by_uuid(cls, uuid: str):
+        return cls.query.filter_by(uuid=uuid)
+
+    @classmethod
+    def find_by_uuid(cls, uuid: str):
+        return cls.query.filter_by(uuid=uuid).first()
+
     def save_to_db(self):
         ModelBase.save_to_db_no_commit(self)
         db.session.commit()
